#1 - Code adjustments.

+++ posmulten-hibernate4-spring-boot/src/test/java/com/github/starnowski/posmulten/demos/NativeSqlWithRlsTest.java
@@ -1,5 +1,6 @@
 package com.github.starnowski.posmulten.demos;
 
+import com.github.starnowski.posmulten.postgresql.core.rls.function.ISetCurrentTenantIdFunctionInvocationFactory;
 import org.assertj.core.api.Assertions;
 import org.junit.Test;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -7,6 +8,7 @@ import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.test.context.jdbc.Sql;
 import org.springframework.test.context.jdbc.SqlConfig;
 
+import static com.github.starnowski.posmulten.demos.TestUtils.statementSettingCurrentTenantVariable;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.springframework.test.context.jdbc.Sql.ExecutionPhase.AFTER_TEST_METHOD;
 import static org.springframework.test.context.jdbc.Sql.ExecutionPhase.BEFORE_TEST_METHOD;
@@ -19,6 +21,9 @@ public class NativeSqlWithRlsTest extends AbstractWebEnvironmentSpringBootTestWi
 
     @Autowired
     protected JdbcTemplate ownerJdbcTemplate;
+
+    @Autowired
+    protected ISetCurrentTenantIdFunctionInvocationFactory setCurrentTenantIdFunctionInvocationFactory;
 
     @Test
     @Sql(value = {TestUtils.CLEAR_DATABASE_SCRIPT_PATH, TestUtils.TEST_BASIC_DATA_SCRIPT_PATH},
@@ -68,7 +73,7 @@ public class NativeSqlWithRlsTest extends AbstractWebEnvironmentSpringBootTestWi
         Assertions.assertThat(TestUtils.countNumberOfRecordsWhere(jdbcTemplate, "user_info", "user_id = 'a0eebc99-9c0b-4ef8-bb6d-6bb9bd380a11' AND tenant_id = 'xds' AND username = 'starnowski'")).isEqualTo(1);
 
         // when
-        jdbcTemplate.execute(statementSettingCurrentTenantVariable("xds") + "UPDATE user_info SET username = 'starnowski1' WHERE user_id = 'a0eebc99-9c0b-4ef8-bb6d-6bb9bd380a11'");
+        jdbcTemplate.execute(statementSettingCurrentTenantVariable(setCurrentTenantIdFunctionInvocationFactory, "xds") + "UPDATE user_info SET username = 'starnowski1' WHERE user_id = 'a0eebc99-9c0b-4ef8-bb6d-6bb9bd380a11'");
 
         // then
         Assertions.assertThat(TestUtils.countNumberOfRecordsWhere(jdbcTemplate, "user_info", "user_id = 'a0eebc99-9c0b-4ef8-bb6d-6bb9bd380a11' AND tenant_id = 'xds' AND username = 'starnowski'")).isZero();
@@ -87,7 +92,7 @@ public class NativeSqlWithRlsTest extends AbstractWebEnvironmentSpringBootTestWi
         Assertions.assertThat(TestUtils.countNumberOfRecordsWhere(jdbcTemplate, "user_info", "user_id = 'a0eebc99-9c0b-4ef8-bb6d-6bb9bd380a11' AND tenant_id = 'xds' AND username = 'starnowski'")).isEqualTo(1);
 
         // when
-        jdbcTemplate.execute(statementSettingCurrentTenantVariable("xds1") + "UPDATE user_info SET username = 'starnowski1' WHERE user_id = 'a0eebc99-9c0b-4ef8-bb6d-6bb9bd380a11'");
+        jdbcTemplate.execute(statementSettingCurrentTenantVariable(setCurrentTenantIdFunctionInvocationFactory, "xds1") + "UPDATE user_info SET username = 'starnowski1' WHERE user_id = 'a0eebc99-9c0b-4ef8-bb6d-6bb9bd380a11'");
 
         // then
         Assertions.assertThat(TestUtils.countNumberOfRecordsWhere(jdbcTemplate, "user_info", "user_id = 'a0eebc99-9c0b-4ef8-bb6d-6bb9bd380a11' AND tenant_id = 'xds' AND username = 'starnowski'")).isEqualTo(1);
@@ -124,7 +129,7 @@ public class NativeSqlWithRlsTest extends AbstractWebEnvironmentSpringBootTestWi
         Assertions.assertThat(TestUtils.countNumberOfRecordsWhere(jdbcTemplate, "user_info", "user_id = 'a0eebc99-9c0b-4ef8-bb6d-6bb9bd380a11' AND tenant_id = 'xds'")).isEqualTo(1);
 
         // when
-        jdbcTemplate.execute(statementSettingCurrentTenantVariable("xds1") + "DELETE FROM user_info WHERE user_id = 'a0eebc99-9c0b-4ef8-bb6d-6bb9bd380a11'");
+        jdbcTemplate.execute(statementSettingCurrentTenantVariable(setCurrentTenantIdFunctionInvocationFactory, "xds1") + "DELETE FROM user_info WHERE user_id = 'a0eebc99-9c0b-4ef8-bb6d-6bb9bd380a11'");
 
         // then
         Assertions.assertThat(TestUtils.countNumberOfRecordsWhere(jdbcTemplate, "user_info", "user_id = 'a0eebc99-9c0b-4ef8-bb6d-6bb9bd380a11' AND tenant_id = 'xds'")).isEqualTo(1);
